The following set of changes include the updating of the README file to reflect what this action is about and its usage.
The addition of a code of conduct and contribution file was added for
developers that would be interested in getting involved with this
project.

Updated the release script such that a user can just include different
information on the command line instead of interactively.

Updated the package information to be inline with this action
implementation.