fix: Incorrectly renamed surreal command on windows.
This set of changes fixes the issue with incorrectly naming the surreal
executable on windows to surrealdb.exe instead of surreal.exe.

As part of this change, the workflows have been updated to determine
that the surreal command line is executed as part of the testing
process.

The package was also updated to version 1.1.0 for the next release.